requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
iniconfig==2.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
Pygments==2.19.1
pytest==8.3.5
requests==2.32.3
rich==14.0.0
urllib3==2.4.0
matplotlib==3.8.4
